Guess sparse mode if restoring to a block device
If the user explicitely specifies "--sparse=always" or "--sparse=never",
this will override any heuristic.

If the restore target supports truncating, it is truncated to zero and
to the final size in turn. We can be sure that sparse mode does the
right thing in this case.

If the restore target does not support truncating, we check the first,
the last, and a random sample of 1% of the blocks. If all of them happen
to be zero, we assume the device has been discarded and use sparse mode
as well.

Fixes #2
Fixes #3