Limit memory usage with LRU cache
We now maintain two caches: One (ro_cache) as LRU read-only cache to
avoid repeated decompression of pages. This other (dirty) is a
traditional HashMap which caches dirty pages indefinitely.

The size of the LRU cache is statically limited (default: 256MiB) while
the dirty page cache could grow up to the size of the underlying image
if all pages were written to. We assume that this is not a normal use
case since the purpose of backy-fuse is to retrieve single files from
backup images.

Thus, it is adivable to mount backy images read-only. Note that even in
this case a limited amount of write operations will take place, e.g. to
update the mounted flag.

Re #7